+++ src/index.ts
@@ -1,4 +1,5 @@
 export { default as TiptopEditor } from './components/editor/TiptopEditor'
+export { TiptopEditorContext, useTiptopEditor } from './components/editor/TiptopEditorContext'
 export * from './types'
 
 import './index.css'